+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to Merlin Spellbook will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+### Added
+- The ability to round to a certain number of decimal places with -round_n notation
+
 ## [0.8.1]
 
 ### Fixed